Add --skipNugetConsistencyCheckPattern command line option
This option takes a wildcard pattern and allows to skip projects matching the pattern when checking NuGet package consistency. This can be useful for test projects, that are not deployed along with other projects, so there is no need to enforce consistency of NuGet packages across projects in a solution.